@@ -3,15 +3,22 @@
 	"version": "0.1.0",
 	"private": true,
 	"homepage": ".",
+	"preinstall": "npx force-resolutions",
 	"dependencies": {
-		"axios": "^0.24.0",
-		"bootstrap": "^4.6.1",
-		"react": "^16.14.0",
-		"react-bootstrap": "^1.6.4",
-		"react-dom": "^16.14.0",
-		"react-router-dom": "^5.3.0",
-		"react-scripts": "3.2.0",
-		"uuid": "^3.4.0"
+		"autoprefixer": "10.4.5",
+		"axios": "^0.27.2",
+		"bootstrap": "^5.1.3",
+		"postcss-preset-env": "^7.5.0",
+		"react": "^18.1.0",
+		"react-bootstrap": "^2.4.0",
+		"react-dom": "^18.1.0",
+		"react-router-dom": "^6.3.0",
+		"react-scripts": "5.0.1",
+		"svgo": "^2.8.0",
+		"uuid": "^8.3.2"
+	},
+	"resolutions": {
+		"autoprefixer": "10.4.5"
 	},
 	"scripts": {
 		"start": "react-scripts start",
@@ -33,5 +40,8 @@
 			"last 1 firefox version",
 			"last 1 safari version"
 		]
+	},
+	"devDependencies": {
+		"force-resolutions": "^1.0.10"
 	}
 }
